Refactor: moves DB codes to it's own module
