+++ server/index.ts
@@ -4,115 +4,21 @@ import { Server } from "socket.io";
 import { faker } from '@faker-js/faker'
 import cors from "cors";
 import { scheduleJob } from 'node-schedule';
-import { PrismaClient, User } from "@prisma/client";
 import { z } from "zod";
 import morgan from 'morgan';
+import { 
+  deleteStaleDBItems,
+  isRoomInDB,
+  addRoomToDB,
+  addUserToRoom,
+  getRoomScore,
+  getUserFromDB,
+} from "./db.js";
 
-const prisma = new PrismaClient()
 
 scheduleJob('*/15 * * * *', async function() {
-  console.log("Deleting old users")
-  // now - (ms * s * mins * hours)
-  let d = new Date(Date.now() - 1000 * 60 * 60 * 16)
-  await prisma.user.deleteMany({
-    where: {
-      lastAccessed : {
-        lte: d
-      }
-    }
-  })
-  console.log("Deleting old rooms")
-  await prisma.room.deleteMany({
-    where: {
-      lastAccessed: {
-        lte: d
-      }
-    }
-  })
+  deleteStaleDBItems()
 })
-
-async function isRoomInDB(roomName: string) {
-  const room = await prisma.room.findUnique({
-    where: {
-      name: roomName
-    }
-  })
-  if (room === null) return false
-  return true
-}
-
-async function addRoomToDB(roomName: string) {
-  await prisma.room.create({
-    data: {
-      name: roomName,
-    }
-  })
-}
-
-async function addUserToRoom(userName: string | User, roomName: string) {
-  // may want to validate if a user exists here
-  if (typeof userName === 'string') {
-    await prisma.room.update({
-      where: {
-        name: roomName
-      },
-      data: {
-        lastAccessed: new Date(),
-        users: {
-          create: {
-            name: userName
-          }
-        }
-      }, 
-    })
-  }
-}
-
-async function getRoomScore(roomName: string) {
-  const roomData = await prisma.room.findUnique({
-    where: {
-      name: roomName
-    },
-    include: {
-      holes: {
-        orderBy: {
-          number: 'asc',
-        }
-      },
-      users: {
-        include: {
-          userScores: true
-        }
-      },
-    },
-  })
-  
-  if (roomData === null) {
-    console.error("Could not fetch room data")
-    return
-  }
-
-  // This Transforms the roomData above into the format for the frontend
-  const players = roomData.users.map(user => {
-    const scores = roomData.holes.map(hole => {
-      const scoreForThisHole = user.userScores.find(userScore => userScore.holeId === hole.id)
-      return scoreForThisHole?.score ?? 0
-    })
-
-   return {
-      name: user.name,
-      id: user.id,
-      scores
-    }
-  })
-
-  const result = {
-    holes: roomData.holes,
-    players,
-  }
-
-  return result
-}
 
 const port = process.env.PORT || 8080
 
@@ -154,11 +60,7 @@ roomIo.on("connect", (socket) => {
       }
 
       // Creates a user and adds them to a room
-      const user = await prisma.user.findFirst({
-        where: {
-          name: validatedUserName
-        }
-      })
+      const user = await getUserFromDB(validatedUserName)
 
       if (!(await isRoomInDB(validatedRoom))) promises.push(addRoomToDB(validatedRoom))
       promises.push(addUserToRoom(user ?? validatedUserName, validatedRoom))
@@ -170,24 +72,6 @@ roomIo.on("connect", (socket) => {
  
     })
     console.log(socket.id)
-
-    socket.on('get-room-state', async ({ roomName, username }, returnDataFn)  => {
-
-      let validatedRoom: string
-      let validatedUserName: string
-
-      try {
-        validatedRoom = z.string().parse(roomName).toLowerCase()
-        validatedUserName = z.string().parse(username)
-      } catch (err) {
-        console.error(err)
-        // onError(JSON.stringify(err, Object.getOwnPropertyNames(err)))
-        return
-      }
-      
-      const score = await getRoomScore(validatedRoom)
-      returnDataFn(score)
-    })
 
     socket.on('update-state', (state) => {
         socket.broadcast.emit("update-state", state)
